moved shared getter/setters into common trait.
