+++ src/Ilios/CoreBundle/Entity/SessionLearningMaterial.php
@@ -3,6 +3,7 @@
 namespace Ilios\CoreBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
+use Ilios\CoreBundle\Traits\LearningMaterialRelationshipEntity;
 use Ilios\CoreBundle\Traits\MeshDescriptorsEntity;
 use Ilios\CoreBundle\Traits\SessionConsolidationEntity;
 use Ilios\CoreBundle\Traits\SortableEntity;
@@ -31,6 +32,7 @@ class SessionLearningMaterial implements SessionLearningMaterialInterface
     use SessionConsolidationEntity;
     use MeshDescriptorsEntity;
     use SortableEntity;
+    use LearningMaterialRelationshipEntity;
 
     /**
      * @var integer
@@ -186,54 +188,6 @@ class SessionLearningMaterial implements SessionLearningMaterialInterface
     }
 
     /**
-     * @param string $notes
-     */
-    public function setNotes($notes)
-    {
-        $this->notes = $notes;
-    }
-
-    /**
-     * @return string
-     */
-    public function getNotes()
-    {
-        return $this->notes;
-    }
-
-    /**
-     * @param boolean $required
-     */
-    public function setRequired($required)
-    {
-        $this->required = $required;
-    }
-
-    /**
-     * @return boolean
-     */
-    public function isRequired()
-    {
-        return $this->required;
-    }
-
-    /**
-     * @param boolean $publicNotes
-     */
-    public function setPublicNotes($publicNotes)
-    {
-        $this->publicNotes = $publicNotes;
-    }
-
-    /**
-     * @return boolean
-     */
-    public function hasPublicNotes()
-    {
-        return $this->publicNotes;
-    }
-
-    /**
      * @param SessionInterface $session
      */
     public function setSession(SessionInterface $session)
@@ -247,53 +201,5 @@ class SessionLearningMaterial implements SessionLearningMaterialInterface
     public function getSession()
     {
         return $this->session;
-    }
-
-    /**
-     * @param LearningMaterialInterface $learningMaterial
-     */
-    public function setLearningMaterial(LearningMaterialInterface $learningMaterial)
-    {
-        $this->learningMaterial = $learningMaterial;
-    }
-
-    /**
-     * @return LearningMaterialInterface
-     */
-    public function getLearningMaterial()
-    {
-        return $this->learningMaterial;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function setStartDate(\DateTime $startDate = null)
-    {
-        $this->startDate = $startDate;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function getStartDate()
-    {
-        return $this->startDate;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function setEndDate(\DateTime $endDate = null)
-    {
-        $this->endDate = $endDate;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function getEndDate()
-    {
-        return $this->endDate;
     }
 }
